requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
certifi==2025.4.26
coverage==7.8.2
gherkin-official==29.0.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
PySocks==1.7.1
pytest==8.3.5
pytest-bdd==8.1.0
pytest-cov==6.1.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
